Passing in arguments scope to the save function
In Railo, maybe ColdFusion as well, the arguments scope to a function is a Struct AND a single key array. I pass the arguments scope as the propertyStructs to the save function. But as it's not an Array in the form we want it, we must check if it's a struct and the make it into an normal array.